2024.4.5a179:
FR-52100 Create the ClientCallbackService in the Web Service api supporting all callback types currently supported in the webhook api with the addition of methods for just displaying information to the user.

Still need to test the new file client callbacks.

Still need to update tutorial here for this important change...