2024.4.10a184:
PR-52136 sapiopylib: ExperimentEntry#entry_status was typed as status enum but is actually received as string when parsed; veloxOnSaveResultMap and veloxOnSaveFieldMapResultMap had key in str instead of int as stated in type.

Furthermore, for ease of use,on WebhookContext, if experiment_entry_list has only a single experiment_entry and experiment_entry in the context is None, then experiment_entry will be the first item in experiment_entry_list. Ditto for data_record in the context.