sapiopylib 2024.08.15a220 NOT COMPATIBLE WITH 24.7
FR-52598:
Add REST and sapiopylib features to create and obtain notebook experiment banner text for ELN experiments.

Add REST and sapioyplib features to
 create, list, and delete notebook experiment tabs in ELN experiments.

Add REST and sapiopylib features to add and remove ELN field definitions of an ELN experiment entry.

Add methods to get default layout of the current user for any data type in data type service.

CR-52604:
Add macro preview since people are complaining about casting into a preview of a date of @today is not working. This cannot be the final value we push to the field map, since the server needs to do the processing and add the server’s time to it. But we can preview it for the purpose of writing business logic in the client.

But also we need to now synchronize after store and commit to ensure the value is updated. We really should probably have done it before, as the value may change after on-save, rules, accessioning, macro, eln formula fields, or other reasons. This however has a performance impact. I am exposing a toggle in record model transaction manager to turn this off as well. But assuming this is turned on, after store and commit, it will update every record model in paged retrieval, and write the latest values to both data record’s fields as well as the current model fields.

PR-52606:
sapiopylib: When obtaining session additional data, overwrite user object’s group name attribute with the current group name obtained from the webservice. This is a valid case for direct webservice call without webhooks.